Configuration-driven tasks; use configured active property
Tasks are now explicitly defined through module configuration.
The names of tasks can also be set explicitly through configuration
and will fallback to fully-qualified class names if not set.

Whether tasks are active is now controlled entirely through
the `active` property of the Task instance and not on the
`active` value from the scheduled_task database table.