Replace ErrorHandler subclass with an ErrorLogTarget to capture and forward errors to the TaskRunner.
Using a subclass of `\yii\console\ErrorHandler` was not a good solution
because it makes it difficult for developers to use other custom
`ErrorHandler` implementations in their own projects.

The solution here is to use a custom log target that will capture errors
logged by the default `ErrorHandler` and forward them to the task runner.